+++ src/test/java/racingcar/domain/CarsTest.java
@@ -1,7 +1,9 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
 import racingcar.moving.MovingStrategy;
 import racingcar.moving.RandomMovingStrategy;
